Only render animations with the same key once
